Rename package http_parser to parser
That was a poor style choice.
