+++ http/src/main/java/org/http4s/blaze/http/parser/BaseExceptions.java
@@ -1,4 +1,4 @@
-package org.http4s.blaze.http.http_parser;
+package org.http4s.blaze.http.parser;
 
 public class BaseExceptions {
