+++ http/src/main/java/org/http4s/blaze/http/parser/Http1ClientParser.java
@@ -1,7 +1,7 @@
-package org.http4s.blaze.http.http_parser;
+package org.http4s.blaze.http.parser;
 
 import java.nio.ByteBuffer;
-import org.http4s.blaze.http.http_parser.BaseExceptions.*;
+import org.http4s.blaze.http.parser.BaseExceptions.*;
 
 public abstract class Http1ClientParser extends BodyAndHeaderParser {
