+++ http/src/main/java/org/http4s/blaze/http/parser/HttpTokens.java
@@ -1,6 +1,6 @@
-package org.http4s.blaze.http.http_parser;
+package org.http4s.blaze.http.parser;
 
-import org.http4s.blaze.http.http_parser.BaseExceptions.BadCharacter;
+import org.http4s.blaze.http.parser.BaseExceptions.BadCharacter;
 
 
 public final class HttpTokens
