+++ http/src/main/java/org/http4s/blaze/http/parser/ParserBase.java
@@ -1,9 +1,9 @@
-package org.http4s.blaze.http.http_parser;
+package org.http4s.blaze.http.parser;
 
 
 import java.nio.ByteBuffer;
-import org.http4s.blaze.http.http_parser.BaseExceptions.BadRequest;
-import org.http4s.blaze.http.http_parser.BaseExceptions.BadCharacter;
+import org.http4s.blaze.http.parser.BaseExceptions.BadRequest;
+import org.http4s.blaze.http.parser.BaseExceptions.BadCharacter;
 
 
 public abstract class ParserBase {
