+++ http/src/test/scala/org/http4s/blaze/http/parser/Benchmarks.scala
@@ -1,4 +1,4 @@
-package org.http4s.blaze.http.http_parser
+package org.http4s.blaze.http.parser
 
 import scala.language.reflectiveCalls
